# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OfferMaker API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/cvs": {
            "post": {
                "tags": [
                    "CVs"
                ],
                "summary": "Upload a new CV",
                "operationId": "926f1b436da2e44d4bcff7948900d31b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "title"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "CV file (PDF, DOC, DOCX, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "title": {
                                        "description": "CV title",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CV uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CV"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cvs/{id}": {
            "get": {
                "tags": [
                    "CVs"
                ],
                "summary": "Get a specific CV",
                "operationId": "8d5a979b0551adc4bb99146f3e6d724a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CV ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CV details with download URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/CV"
                                        },
                                        {
                                            "properties": {
                                                "download_url": {
                                                    "type": "string",
                                                    "format": "url"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "CV not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "CVs"
                ],
                "summary": "Update a CV",
                "operationId": "c15d7b0ac74fab4ebf5fe895c9a70228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CV ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "CV file (PDF, DOC, DOCX, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "title": {
                                        "description": "CV title",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CV updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CV"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "CV not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CVs"
                ],
                "summary": "Delete a CV",
                "operationId": "4be7c7e58f10d37eb9c4cdb9c323a905",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "CV ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CV deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "CV not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CV": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "file_path": {
                        "type": "string"
                    },
                    "file_type": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}